Snail/test_solution.py
import pytest
from solution import snail

test_data = [
    ([[1, 2, 3], [4, 5, 6], [7, 8, 9]], [1, 2, 3, 6, 9, 8, 7, 4, 5]),
    ([[1, 2, 3], [8, 9, 4], [7, 6, 5]], [1, 2, 3, 4, 5, 6, 7, 8, 9]),
    (
        [
            [1, 2, 3, 4, 5],
            [6, 7, 8, 9, 10],
            [11, 12, 13, 14, 15],
            [16, 17, 18, 19, 20],
            [21, 22, 23, 24, 25]
        ],
        [1, 2, 3, 4, 5, 10, 15, 20, 25, 24, 23, 22, 21, 16, 11, 6, 7, 8, 9, 14, 19, 18, 17, 12, 13]
    )
]


def test_func():
    example = [[]]
    snail(example)


def test_func_output():
    example = [[]]
    resp = snail(example)
    assert list == type(resp)


@pytest.mark.parametrize(['example', 'expected'], test_data)
def test_solution(example, expected):
    resp = snail(example)
    assert expected == resp
